PXComponentAction moved into a new file
